Only update the queue for the next bucket if data has been aggregated
Data in a bucket is aggregated only when the full time slice for that
data has passed. This commit introduces logic that updates the queue for
the next bucket only when data is aggregated.

Here is an example to illustrate. Suppose we finish aggregating data
data in the one hour bucket. That is, one hour data is rolled up into
the 6 hour bucket. Next we add an entry to the 6 hour queue for the data
that was just aggregated. Next we start working on the 6 hour data.
Suppose we are in the middle of the time slice for the current data.
Since we wait until the 6 hour time slice has passed before rolling that
data up into the 24 hour bucket, we do not add an entry in the 24 hour
queue (for that data).

This commit also add support for "removing" entries from the queues.
Now when data from a queue entry is rolled up, that entry is deleted
from the queue.