updating schema script to define trait column families
Two new column families are added in this commit. The first is for trait
data and the second is an index for resource trait queries.

Also updating pom.xml to include missing runtime dependencies in the
plugin jar file. Updating the plugin descriptor with configuration
properties.

Lastly, there are some changes to simplify the database purge code in
tests. Now there is a deleteAllRows method that does not require any row
keys to be specified.