Create bundle types only if necessary  for create local cluster operation
Deploying a bundle involves several steps including creating the bundle,
creating the bundle version, creating the bundle destination, and lastly
the bundle deployment. Everything except for the deployment only needs
to be created once. Logic has been added to see if each of those parts
exist. If not, they are created and used.