First cut at implementing addNumericDataMethod
The addNumericData method is responsible for persisting raw numeric
metric data. This implementation (so far) does a couple things. First,
it stores the raw data into a cache. Then a task separate task is kicked
off to insert or update a separate entry in another cache. This entry
stores a set of raw data. The raw data is batched by the hour.

This commit includes a unit test to verify the aforementioned behavior.
It is an initial test that only covers a basic case. More tests are
needed for things like handling multiple schedule ids, handling data
across the hourly boundaries, and more.