Insert entry into hourly index cache when raw data is inserted
This commit introduces an hourly data index cache that is updated
whenever raw data is inserted. The purpose of this index caches is to
make it quick and effecient to look up and determine the set of input
keys for compression/aggregation.

This is also the initial commit for RawDataBatch. It stores a set of raw
data where each entry is grouped by hour. It encapsulates the set that
was previously used as the cache entry. That set has also been
refactored into a SortedSet where the raw data is ordered by timestamp.