First pass at calculateAggregates() method
This commit adds initial support for aggregating raw data. There is no
support yet for handling 1hr, 6hr, or 24hr aggregates. The approach is
as follows. For the bucket that data is going into we will have
a separate cache for each time slice interval in that bucket. For one
hour data there will be 24 caches (though this may later reduced to 12),
for 6 hr data, there will be 4 such caches. Each of these caches serves
as an index to make it easier/faster to find metric data that need
to be compressed. The new map reduce job that consists of
GetMetricKeysMapper and GetMetricKeysReducer handles querying the
appropriate index cache to get the set of metric data to be compressed.
Then we pass an instance of UpdateMetricDataBactch to
DistributedExecutorService to finish the work of batching up the data.